requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
redis==5.0.4
pytrends==4.9.2
google-ads==27.0.0
python-dotenv==1.0.1
pydantic==2.7.4
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
locust==2.29.1